Split mc_measure_set.hpp into multiple header files and format them
